@@ -1,20 +1,14 @@
 name: app
 description: A sample command-line application.
-author: axel-op <49279289+axel-op@users.noreply.github.com>
-# version: 1.0.0
-# homepage: https://www.example.com
 
 environment:
-  sdk: ">=2.6.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
-  args: ^1.5.2
-  meta: ^1.1.7
-  path: ^1.6.4
-  github: ^6.2.0
-  # logging: ^0.11.4
-  github_actions_toolkit: 0.0.3
+  args: ^2.4.1
+  github: ^9.14.0
+  github_actions_toolkit: ^0.0.8
+  path: ^1.8.3
 
 dev_dependencies:
-  pedantic: ^1.8.0
-  extra_pedantic: any
+  lints: ^2.0.1
