@@ -2,25 +2,18 @@ import 'dart:convert';
 import 'dart:io';
 
 import 'package:app/paths.dart';
-import 'package:github/github.dart';
 import 'package:github_actions_toolkit/github_actions_toolkit.dart';
-import 'package:meta/meta.dart';
 
 const Input githubTokenInput = Input(
-  'githubToken',
-  isRequired: true,
-  canBeEmpty: false,
-),
+      'githubToken',
+      isRequired: true,
+      canBeEmpty: false,
+    ),
     packagePathInput = Input(
-  'relativePath',
-  isRequired: false,
-  canBeEmpty: true,
-),
-    minAnnotationLevelInput = Input(
-  'minAnnotationLevel',
-  isRequired: true,
-  canBeEmpty: false,
-);
+      'relativePath',
+      isRequired: false,
+      canBeEmpty: true,
+    );
 
 class Inputs {
   /// Token to call the GitHub API
@@ -31,9 +24,6 @@ class Inputs {
 
   /// Slug of the repository
   final String repositorySlug;
-
-  /// Minimum level of the diff annotations
-  final CheckRunAnnotationLevel minAnnotationLevel;
 
   final Paths paths;
 
@@ -49,28 +39,26 @@ class Inputs {
 
     return Inputs._(
       commitSha: _sha,
-      githubToken: githubTokenInput.value,
-      minAnnotationLevel: _minAnnotationLevel,
+      githubToken: githubTokenInput.value!,
       paths: paths,
-      repositorySlug: Platform.environment['GITHUB_REPOSITORY'],
+      repositorySlug: Platform.environment['GITHUB_REPOSITORY']!,
     );
   }
 
   Inputs._({
-    @required this.commitSha,
-    @required this.githubToken,
-    @required this.minAnnotationLevel,
-    @required this.paths,
-    @required this.repositorySlug,
+    required this.commitSha,
+    required this.githubToken,
+    required this.paths,
+    required this.repositorySlug,
   });
 
   static String get _sha {
-    final String pathEventPayload = Platform.environment['GITHUB_EVENT_PATH'];
+    final String pathEventPayload = Platform.environment['GITHUB_EVENT_PATH']!;
     final Map<String, dynamic> eventPayload =
         jsonDecode(File(pathEventPayload).readAsStringSync());
-    final String commitSha = Platform.environment['GITHUB_SHA'];
+    final String commitSha = Platform.environment['GITHUB_SHA']!;
     stderr.writeln('SHA that triggered the workflow: $commitSha');
-    final Map<String, dynamic> pullRequest = eventPayload['pull_request'];
+    final Map<String, dynamic>? pullRequest = eventPayload['pull_request'];
     if (pullRequest != null) {
       final String baseSha = pullRequest['base']['sha'];
       final String headSha = pullRequest['head']['sha'];
@@ -81,20 +69,5 @@ class Inputs {
       }
     }
     return commitSha;
-  }
-
-  static CheckRunAnnotationLevel get _minAnnotationLevel {
-    const Map<String, CheckRunAnnotationLevel> annotationMapping = {
-      'info': CheckRunAnnotationLevel.notice,
-      'warning': CheckRunAnnotationLevel.warning,
-      'error': CheckRunAnnotationLevel.failure,
-    };
-    final CheckRunAnnotationLevel level =
-        annotationMapping[minAnnotationLevelInput.value.toLowerCase()];
-    if (level == null) {
-      throw ArgumentError.value(
-          minAnnotationLevelInput.value, 'minAnnotationLevel');
-    }
-    return level;
   }
 }
