@@ -1,12 +1,11 @@
 import 'dart:io';
 
-import 'package:meta/meta.dart';
 import 'package:path/path.dart' as path;
 
 class Paths {
   final String packageRelativePath;
 
-  Paths({@required this.packageRelativePath});
+  Paths({required this.packageRelativePath});
 
   /// Canonical path to the package to analyze
   String get canonicalPathToPackage =>
@@ -15,7 +14,7 @@ class Paths {
   /// Path to the folder containing the entire repository
   String get canonicalPathToRepoRoot {
     const String envVarWorkspace = 'GITHUB_WORKSPACE';
-    final String repoPath = Platform.environment[envVarWorkspace];
+    final String? repoPath = Platform.environment[envVarWorkspace];
     if (repoPath == null) {
       throw ArgumentError.value(repoPath, envVarWorkspace,
           "Make sure you call 'actions/checkout' in a previous step. Invalid environment variable");
