@@ -3,7 +3,7 @@ import 'dart:io';
 
 import 'package:app/github.dart';
 import 'package:app/inputs.dart';
-import 'package:app/pana_result.dart';
+import 'package:app/report.dart';
 import 'package:github_actions_toolkit/github_actions_toolkit.dart' as gaction;
 
 const logger = gaction.log;
@@ -28,7 +28,7 @@ dynamic main(List<String> args) async {
     }
   }
 
-  Future<void> _exitProgram([dynamic cause]) async {
+  Future<void> exitProgram([dynamic cause]) async {
     await tryCancelAnalysis(cause);
     await Future.wait<dynamic>([stderr.done, stdout.done]);
     logger.error('Exiting with code $exitCode');
@@ -36,10 +36,16 @@ dynamic main(List<String> args) async {
   }
 
   try {
+    final canonicalPathToRepoRoot = inputs.paths.canonicalPathToRepoRoot;
+    final userProcessResult = await gaction.exec('whoami', [], silent: true);
+    final user = (userProcessResult.stdout as String).trim();
+    logger.info('whoami returned: $user');
+    await gaction.exec('chown', [user, '-R', canonicalPathToRepoRoot]);
+
     // Command to disable analytics reporting, and also to prevent a warning from the next command due to Flutter welcome screen
     await logger.group(
       'Disabling Flutter analytics',
-      () => gaction.exec('flutter', const <String>['config', '--no-analytics']),
+      () => gaction.exec('flutter', const ['config', '--no-analytics']),
     );
 
     await analysis.start();
@@ -48,20 +54,19 @@ dynamic main(List<String> args) async {
     logger.startGroup('Running pana');
     final panaProcessResult = await gaction.exec(
       'pana',
-      <String>[
-        '--scores',
-        '--no-warning',
-        '--source',
-        'path',
-        inputs.paths.canonicalPathToPackage,
-      ],
+      ['--json', '--no-warning', inputs.paths.canonicalPathToPackage],
     );
     logger.endGroup();
 
     if (panaProcessResult.exitCode != 0) {
       logger.error('Pana exited with code ${panaProcessResult.exitCode}');
       exitCode = panaProcessResult.exitCode;
-      await _exitProgram();
+      await exitProgram();
+    }
+    if (panaProcessResult.stderr
+        .toLowerCase()
+        .contains("can't load kernel binary")) {
+      throw Exception("SDK incompatibility");
     }
     if (panaProcessResult.stdout == null) {
       throw Exception('The pana command has returned no valid output.'
@@ -69,18 +74,17 @@ dynamic main(List<String> args) async {
           ' Please file an issue at https://github.com/axel-op/dart-package-analyzer/issues/new');
     }
 
-    final panaResult = PanaResult.fromOutput(
-      jsonDecode(panaProcessResult.stdout) as Map<String, dynamic>,
-      paths: inputs.paths,
-    );
+    final report = Report.fromOutput(
+        jsonDecode(panaProcessResult.stdout) as Map<String, dynamic>);
+
+    if (report.errorMessage != null) {
+      throw Exception(report.errorMessage);
+    }
 
     // Posting comments on GitHub
     await logger.group(
       'Publishing report',
-      () async => analysis.complete(
-        panaResult: panaResult,
-        minAnnotationLevel: inputs.minAnnotationLevel,
-      ),
+      () async => analysis.complete(report: report),
     );
 
     // Setting outputs
@@ -88,12 +92,23 @@ dynamic main(List<String> args) async {
       'Setting outputs',
       () async {
         final outputs = <String, String>{
-          'health': panaResult.healthScore.toStringAsFixed(2),
-          'maintenance': panaResult.maintenanceScore.toStringAsFixed(2),
-          'errors': panaResult.analyzerResult.errorCount.toString(),
-          'warnings': panaResult.analyzerResult.warningCount.toString(),
-          'hints': panaResult.analyzerResult.hintCount.toString()
+          "json_output": jsonEncode(jsonDecode(panaProcessResult.stdout)),
+          "total": report.grantedPoints.toString(),
+          "total_max": report.maxPoints.toString()
         };
+        final idsToKeys = <String, String>{
+          "convention": "conventions",
+          "documentation": "documentation",
+          "platform": "platforms",
+          "analysis": "analysis",
+          "dependency": "dependencies",
+          "null-safety": "null_safety",
+        };
+        for (final section in report.sections) {
+          final key = idsToKeys[section.id] ?? section.id;
+          outputs[key] = section.grantedPoints.toString();
+          outputs["${key}_max"] = section.maxPoints.toString();
+        }
         for (final output in outputs.entries) {
           logger.info('${output.key}: ${output.value}');
           gaction.setOutput(output.key, output.value);
@@ -108,6 +123,5 @@ dynamic main(List<String> args) async {
 }
 
 void _writeError(dynamic error, StackTrace stackTrace) {
-  logger.error(error.toString() +
-      (stackTrace != null ? '\n' + stackTrace.toString() : ''));
+  logger.error('$error\n$stackTrace');
 }
